zsdl: add sdl_ttf bindings
