+++ libs/zsdl/libs/macos/Frameworks/SDL2_ttf.framework/Headers
@@ -1,0 +1,1 @@
+Versions/Current/Headers
