+++ libs/zsdl/libs/macos/Frameworks/SDL2_ttf.framework/SDL2_ttf
@@ -1,0 +1,1 @@
+Versions/Current/SDL2_ttf
