+++ libs/zsdl/libs/macos/Frameworks/SDL2_ttf.framework/Versions/A/SDL2_ttf
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:c6d077882fdcbcd52ba521dac60608799c8e7910535435cf449506d17f1d3ef0
+size 3371536
