+++ libs/zsdl/libs/macos/Frameworks/SDL2_ttf.framework/Versions/A/_CodeSignature/CodeResources
@@ -1,0 +1,233 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>files</key>
+	<dict>
+		<key>Resources/CMake/sdl2_ttf-config-version.cmake</key>
+		<data>
+		2H0oVk3V+kXlw2xJrtkplR8uzZk=
+		</data>
+		<key>Resources/CMake/sdl2_ttf-config.cmake</key>
+		<data>
+		BnA2e1CmuuadQ4WmodFyIid1EAo=
+		</data>
+		<key>Resources/FTL.TXT</key>
+		<data>
+		aTfQN2a6wPh5l5DbBXHTetZHhw8=
+		</data>
+		<key>Resources/FreeType-LICENSE.txt</key>
+		<data>
+		ZLfyE93XJpXZSGahqVMu5bOkcqg=
+		</data>
+		<key>Resources/HarfBuzz-LICENSE.txt</key>
+		<data>
+		GMGU+yuWtqYCiaeSZedpdv/bMD0=
+		</data>
+		<key>Resources/Info.plist</key>
+		<data>
+		OOVyvVBcUvPbIq/C1ppKHIYu7JM=
+		</data>
+		<key>Resources/ReadMe.txt</key>
+		<data>
+		6McVk9LCuB72GPNUFfitazzPTSs=
+		</data>
+	</dict>
+	<key>files2</key>
+	<dict>
+		<key>Headers/SDL_ttf.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			4+ZefqoD9HA5z8lHT5i3lh6GLrI=
+			</data>
+			<key>hash2</key>
+			<data>
+			RtPoc6u/bhG6XOP0JMhPf8KuxCY/61Cf4kSU0ktp5J0=
+			</data>
+		</dict>
+		<key>Resources/CMake/sdl2_ttf-config-version.cmake</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			2H0oVk3V+kXlw2xJrtkplR8uzZk=
+			</data>
+			<key>hash2</key>
+			<data>
+			PfjH4G+WE/+w6VOYm4THELt28ySVkCpaiqk0eqEWK2w=
+			</data>
+		</dict>
+		<key>Resources/CMake/sdl2_ttf-config.cmake</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			BnA2e1CmuuadQ4WmodFyIid1EAo=
+			</data>
+			<key>hash2</key>
+			<data>
+			UXGxuOqwWrvD/blIrqL0Y58GXloOCxbZzGzoWI/51QE=
+			</data>
+		</dict>
+		<key>Resources/FTL.TXT</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			aTfQN2a6wPh5l5DbBXHTetZHhw8=
+			</data>
+			<key>hash2</key>
+			<data>
+			CME1dV3ViQOUcPH9u0ANqquqpQ0LNm0Zzr/00imGuqE=
+			</data>
+		</dict>
+		<key>Resources/FreeType-LICENSE.txt</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			ZLfyE93XJpXZSGahqVMu5bOkcqg=
+			</data>
+			<key>hash2</key>
+			<data>
+			/QVt5BlpA6Z2II71jP3a/H1YPR8o+i5EwwnPhKWeYvs=
+			</data>
+		</dict>
+		<key>Resources/HarfBuzz-LICENSE.txt</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			GMGU+yuWtqYCiaeSZedpdv/bMD0=
+			</data>
+			<key>hash2</key>
+			<data>
+			vrPwF860E4dmf3gjenzs+E3voO0lfFBYlGY9ap8TIRs=
+			</data>
+		</dict>
+		<key>Resources/Info.plist</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			OOVyvVBcUvPbIq/C1ppKHIYu7JM=
+			</data>
+			<key>hash2</key>
+			<data>
+			A+LX07DOgTlN2hjyvdv4MWJQQMamT82SVeURsGCJ+eI=
+			</data>
+		</dict>
+		<key>Resources/ReadMe.txt</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			6McVk9LCuB72GPNUFfitazzPTSs=
+			</data>
+			<key>hash2</key>
+			<data>
+			qiHN8zxBBJMv8RKGQFiuwXPpgMW7quHFipUT1qlq2wk=
+			</data>
+		</dict>
+	</dict>
+	<key>rules</key>
+	<dict>
+		<key>^Resources/</key>
+		<true/>
+		<key>^Resources/.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^Resources/.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Resources/Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^version.plist$</key>
+		<true/>
+	</dict>
+	<key>rules2</key>
+	<dict>
+		<key>.*\.dSYM($|/)</key>
+		<dict>
+			<key>weight</key>
+			<real>11</real>
+		</dict>
+		<key>^(.*/)?\.DS_Store$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>2000</real>
+		</dict>
+		<key>^(Frameworks|SharedFrameworks|PlugIns|Plug-ins|XPCServices|Helpers|MacOS|Library/(Automator|Spotlight|LoginItems))/</key>
+		<dict>
+			<key>nested</key>
+			<true/>
+			<key>weight</key>
+			<real>10</real>
+		</dict>
+		<key>^.*</key>
+		<true/>
+		<key>^Info\.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^PkgInfo$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^Resources/</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^Resources/.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^Resources/.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Resources/Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^[^/]+$</key>
+		<dict>
+			<key>nested</key>
+			<true/>
+			<key>weight</key>
+			<real>10</real>
+		</dict>
+		<key>^embedded\.provisionprofile$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^version\.plist$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+	</dict>
+</dict>
+</plist>
