+++ libs/zsdl/libs/macos/Frameworks/SDL2_ttf.framework/Versions/Current
@@ -1,0 +1,1 @@
+A
