+++ libs/zsdl/libs/x86_64-linux-gnu/lib/libSDL2_ttf-2.0.so
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:cb1e7c5b8211613b63f542a2d6d30925c95daa7883f854ed736950895bd75406
+size 82384
