+++ libs/zsdl/libs/x86_64-windows-gnu/bin/SDL2_ttf.dll
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:d96cb0005270afe9925c0bde9c95e29dd7a3f287f79aea17ca381fe3daa28e0e
+size 1561088
